+++ core/include/webview.h
@@ -1,7 +1,7 @@
 /**
  * @file webview.h
  *
- * @deprecated This header file is deprecated. Use `webview/core/webview.h` instead.
+ * @deprecated This header file is deprecated. Use `webview/webview.h` instead.
  *
  * This file is provided for backward-compatibility with existing code
  * such as `#include "webview.h"`.
@@ -10,6 +10,6 @@
 #ifndef WEBVIEW_ROOT_H
 #define WEBVIEW_ROOT_H
 
-#include "webview/core/webview.h"
+#include "webview/webview.h"
 
 #endif /* WEBVIEW_ROOT_H */
