+++ examples/c/bind/src/main.c
@@ -1,4 +1,4 @@
-#include "webview/core/webview.h"
+#include "webview/webview.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
