+++ examples/cc/basic/src/main.cc
@@ -1,4 +1,4 @@
-#include "webview/core/webview.h"
+#include "webview/webview.h"
 
 #include <iostream>
